requirements.txt
networkx
pyvis
pymupdf
tqdm
python-dotenv
openai
Pillow
